perf: defer timeline annotation chrome
